New user screens with menu build-out setup
